# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: AIAE Operational Hub API
  version: "1"
  description: >
    Backend API for AI Digital Operational Hub RBAC foundation. Authentication is Clerk SSO;
    the backend validates Clerk-issued JWTs against Clerk JWKS. OpenAPI is the source of truth
    for all exposed REST endpoints.
servers:
  - url: /
    description: Endpoint paths are fully qualified under /api/v1.
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/campaigns/search:
    post:
      tags:
        - campaigns
      operationId: searchCampaigns
      description: >
        Searches campaigns visible to the authenticated user, sourced from BigQuery.
        Optionally filtered by client_id or agency_id. Sorting and paging are applied server-side.
      parameters:
        - $ref: '#/components/parameters/PageNumberParam'
        - $ref: '#/components/parameters/PageSizeParam'
      requestBody:
        description: Filter and sort criteria for the campaign search.
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignSearchRequestV1'
      responses:
        '200':
          description: A page of campaigns was resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignPageResponseV1'
        '400':
          description: The paging parameters or search criteria failed validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while reading campaigns from BigQuery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/campaigns/{campaignId}:
    get:
      tags:
        - campaigns
      operationId: getCampaign
      description: >
        Returns one campaign by id, sourced from the netsuite_campaigns_with_ids_fresh_data BigQuery
        table, enforcing the current user's agency visibility (a campaign outside it resolves the same
        as an unknown one). Lets the campaign workspace be opened directly - by a pasted link, a
        bookmark, or a restored tab - instead of only from a list that carries the campaign along.
      parameters:
        - name: campaignId
          in: path
          required: true
          description: The campaign id.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The campaign was resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '404':
          description: The campaign was not found, or is not visible to the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while reading the campaign from BigQuery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/campaigns/{campaignId}/report-rows:
    post:
      tags:
        - campaigns
      operationId: listCampaignReportRows
      description: >
        Returns one page of per-day, per-line-item delivery/actuals rows for one campaign, sourced from
        the platform_mart_adjustments_view_op_hub BigQuery view (merges base platform_mart delivery
        data with manual op-hub adjustments; conflicts resolve by latest last_modified_at
        server-side). Metric totals and the date range / distinct line item count are always computed
        over every matching row that matches the requested filters, not just the returned page, so they
        stay stable while more pages load. POST (not GET) because the filter set is a variable-length
        list of multi-value dimension filters - the same shape /campaigns/search already uses.
      parameters:
        - name: campaignId
          in: path
          required: true
          description: The campaign id.
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageNumberParam'
        - $ref: '#/components/parameters/PageSizeParam'
      requestBody:
        description: Filter and sort criteria for the report rows.
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRowSearchRequestV1'
      responses:
        '200':
          description: A page of the campaign's report rows was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportRowsPageResponseV1'
        '400':
          description: The paging parameters or filter/sort criteria failed validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '404':
          description: The campaign was not found, or is not visible to the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while reading report rows from BigQuery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/campaigns/{campaignId}/report-rows/distinct-values:
    get:
      tags:
        - campaigns
      operationId: listReportRowDistinctValues
      description: >
        Returns the distinct, non-null values of one report-rows dimension for one campaign, sourced
        from the same platform_mart_adjustments_view_op_hub rows the report-rows endpoint reads - used
        to populate a column's filter picker. Capped at 500 values; metrics are not filterable (see
        ReportRowFilterFieldEnumV1).
      parameters:
        - name: campaignId
          in: path
          required: true
          description: The campaign id.
          schema:
            type: integer
            format: int64
        - name: field
          in: query
          required: true
          description: The dimension to list distinct values for.
          schema:
            $ref: '#/components/schemas/ReportRowFilterFieldEnumV1'
      responses:
        '200':
          description: The dimension's distinct values were found.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: The field parameter failed validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '404':
          description: The campaign was not found, or is not visible to the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while reading distinct values from BigQuery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/campaigns/{campaignId}/report-rows/adjustments:
    post:
      tags:
        - campaigns
      operationId: saveReportRowAdjustments
      description: >
        Appends one or more report-row adjustments (inline metric edits and/or manually-added rows) for
        one campaign - each is written as a new, append-only row to the adjustments write table, stamped
        with the current user and a server-authoritative timestamp, so the report-rows read view merges
        it over the base data (latest last-modified timestamp wins). A single inline cell edit is a
        batch of one; a bulk adjustment is a batch of many - both use this same endpoint.
      parameters:
        - name: campaignId
          in: path
          required: true
          description: The campaign id.
          schema:
            type: integer
            format: int64
      requestBody:
        description: The adjustments to append.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRowAdjustmentsRequestV1'
      responses:
        '204':
          description: The adjustments were written.
        '400':
          description: The adjustments payload failed validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '404':
          description: The campaign was not found, or is not visible to the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while writing adjustments to BigQuery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
...
  /api/v1/campaigns/{campaignId}/insertion-orders:
    get:
      tags:
        - campaigns
      operationId: listCampaignInsertionOrders
      description: >
        Returns one campaign's insertion orders (US-009…011), each with its real NetSuite line items,
        sourced from the netsuite_campaigns_with_ids_fresh_data IO Lines table. One insertion order per
        order_id; one line item per row of that table. Order-level values (budget/status/dates) are
        read once per order, never summed or repeated across its line items. The source has no
        per-line-item status - a line item's status is its order's. Not paginated: a campaign's IO/LI
        tree is small and rendered whole.
      parameters:
        - name: campaignId
          in: path
          required: true
          description: The campaign id.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The campaign's insertion orders were returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InsertionOrderV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '404':
          description: The campaign was not found, or is not visible to the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while reading insertion orders from BigQuery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT bearer token validated against Clerk JWKS.
  parameters:
    PageNumberParam:
      name: pageNumber
      in: query
      required: false
      description: One-based page number to return.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      description: Number of items per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    CampaignSortFieldV1:
      type: object
      description: Single-field sort directive for the campaign search.
      properties:
        field:
          $ref: '#/components/schemas/CampaignSortFieldEnumV1'
        direction:
          $ref: '#/components/schemas/DirectionEnumV1'
      required:
        - field
...
    CampaignSearchRequestV1:
      type: object
      description: Filter and sort criteria for searching campaigns.
      properties:
        filters:
          type: array
          description: >
            Filters applied additively (AND) to the campaign search, with one exception: several
            AGENCY_ID filters are OR-ed into a single IN, so a caller can filter by a set of agencies
            at once. (ANDing equality on the same id column would match nothing, so OR is the only
            sensible reading.) This never widens visibility - the resulting predicate still ANDs with
            the caller's own agency visibility, so the effective set is the intersection of the two.
          items:
            $ref: '#/components/schemas/CampaignFilterFieldV1'
        sorting:
          $ref: '#/components/schemas/CampaignSortFieldV1'
...